requirements.txt
fastapi>=0.111.0
uvicorn>=0.28.0
sqlalchemy>=2.0.30
python-jose[cryptography]>=3.3.0
pydantic>=2.7.4